requirements.txt
Django==3.1.1
channels==2.4.0
djangorestframework==3.11.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
channels-redis==3.1.0
django-heroku==0.3.1